[MISC, DOC] Move tests and snippets
