+++ test/performance/alphabet/CMakeLists.txt
@@ -1,3 +1,5 @@
+add_subdirectories ()
+
 seqan3_benchmark(alphabet_assign_char_benchmark.cpp)
 seqan3_benchmark(alphabet_assign_rank_benchmark.cpp)
 seqan3_benchmark(alphabet_to_char_benchmark.cpp)
