+++ test/performance/range/views/CMakeLists.txt
@@ -3,6 +3,3 @@ seqan3_benchmark(view_drop_benchmark.cpp)
 seqan3_benchmark(view_drop_view_take_benchmark.cpp)
 seqan3_benchmark(view_take_benchmark.cpp)
 seqan3_benchmark(view_take_until_benchmark.cpp)
-seqan3_benchmark(view_translate_1D_benchmark.cpp)
-seqan3_benchmark(view_translate_2D_benchmark.cpp)
-seqan3_benchmark(view_translate_2D_1D_benchmark.cpp)
