+++ test/snippet/alphabet/views/range_view_to_char.cpp
