+++ test/snippet/alphabet/views/rank_to.cpp
