+++ test/snippet/alphabet/views/to_char.cpp
