+++ test/snippet/alphabet/views/to_rank.cpp
