+++ test/snippet/alphabet/views/translate_dna5.cpp
