+++ test/snippet/alphabet/views/translate_join.cpp
