+++ test/snippet/alphabet/views/translate_usage.cpp
