+++ test/snippet/alphabet/views/trim_quality_dna5q.cpp
