+++ test/snippet/alphabet/views/trim_quality_phred42.cpp
