+++ test/unit/alphabet/views/CMakeLists.txt
@@ -1,3 +1,10 @@
 add_subdirectories()
 
-seqan3_test (complement_test.cpp)
+seqan3_test(char_to_test.cpp)
+seqan3_test(complement_test.cpp)
+seqan3_test(rank_to_test.cpp)
+seqan3_test(to_char_test.cpp)
+seqan3_test(to_rank_test.cpp)
+seqan3_test(translate_join_test.cpp)
+seqan3_test(translate_test.cpp)
+seqan3_test(trim_quality_test.cpp)
