+++ test/unit/alphabet/views/trim_quality_test.cpp
