+++ test/unit/range/views/CMakeLists.txt
@@ -2,21 +2,14 @@ add_subdirectories()
 
 seqan3_test(adaptor_base_test.cpp)
 seqan3_test(as_const_test.cpp)
-seqan3_test(char_to_test.cpp)
 seqan3_test(convert_test.cpp)
 seqan3_test(deep_test.cpp)
 seqan3_test(drop_test.cpp)
 seqan3_test(istreambuf_test.cpp)
 seqan3_test(move_test.cpp)
-seqan3_test(rank_to_test.cpp)
 seqan3_test(type_reduce_test.cpp)
 seqan3_test(take_test.cpp)
 seqan3_test(take_until_test.cpp)
 seqan3_test(take_line_test.cpp)
-seqan3_test(to_char_test.cpp)
 seqan3_test(to_lower_test.cpp)
-seqan3_test(to_rank_test.cpp)
 seqan3_test(to_upper_test.cpp)
-seqan3_test(translate_join_test.cpp)
-seqan3_test(translate_test.cpp)
-seqan3_test(trim_quality_test.cpp)
